requirements.txt
numpy
zope.interface==4.1.1
Twisted==13.1
nose==1.3.7